New ‘--no-output’ options to ‘fdrdf’.
src/fdrdf.pl: New ‘--no-output’ option added to the
‘Getopt::Mixed::init’ argument and is now handled by the arguments'
parsing loop code; do not serialize the model if ‘--no-output’ is given.
(help): Document the new ‘--no-output’ option.
($out): Check that ‘$output’ is not ‘undef’, don't open the file and
hold ‘undef’ if it is.
doc/fdrdf.xml (#options): Document the new ‘--no-output’ option.